[HOPS-1367] fix locking in addStoredBlockTx
