+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/io/hops/transaction/lock/BaseINodeLock.java
@@ -15,17 +15,25 @@
  */
 package io.hops.transaction.lock;
 
+import com.google.common.base.Joiner;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
+import io.hops.common.INodeResolver;
+import io.hops.common.INodeUtil;
 import io.hops.exception.StorageException;
 import io.hops.exception.TransactionContextException;
+import io.hops.exception.TransientStorageException;
 import io.hops.metadata.hdfs.dal.INodeDataAccess;
 import io.hops.resolvingcache.Cache;
 import io.hops.metadata.hdfs.entity.INodeCandidatePrimaryKey;
+import io.hops.metadata.hdfs.entity.INodeIdentifier;
 import io.hops.transaction.EntityManager;
+import static io.hops.transaction.lock.Lock.LOG;
+import java.io.IOException;
 import org.apache.hadoop.hdfs.server.namenode.*;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
@@ -33,6 +41,9 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
+import org.apache.commons.math3.stat.StatUtils;
+import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.hdfs.protocol.UnresolvedPathException;
 import org.apache.hadoop.hdfs.server.namenode.INodeDirectory;
 
 public abstract class BaseINodeLock extends Lock {
@@ -353,5 +364,387 @@ public abstract class BaseINodeLock extends Lock {
     }else{
       return TransactionLockTypes.INodeLockType.READ_COMMITTED;
     }
+  }
+
+  protected abstract class CacheResolver {
+
+    abstract List<INode> fetchINodes(final TransactionLockTypes.INodeLockType lockType, String path, boolean resolveLink)
+        throws IOException;
+
+    abstract List<INode> fetchINodes(final TransactionLockTypes.INodeLockType lockType, long inodeId) throws IOException;
+
+    protected int verifyINodesFull(final List<INode> inodes, final String[] names, final long[] parentIds,
+        final long[] inodeIds) throws IOException {
+      int index = -1;
+      if (names.length == parentIds.length) {
+        if (inodes.size() == names.length) {
+          index = verifyINodesPartial(inodes, names, parentIds, inodeIds);
+        }
+      }
+      return index;
+    }
+
+    protected int verifyINodesPartial(final List<INode> inodes, final String[] names, final long[] parentIds,
+        final long[] inodeIds) throws IOException {
+      int index = (int) StatUtils.min(new double[]{inodes.size(), inodeIds.length, parentIds.length, names.length});
+      for (int i = 0; i < index; i++) {
+        INode inode = inodes.get(i);
+        boolean noChangeInInodes = inode != null && inode.getLocalName().equals(names[i]) && inode.getParentId()
+            == parentIds[i] && inode.getId() == inodeIds[i];
+        if (!noChangeInInodes) {
+          index = i;
+          break;
+        }
+      }
+      return index;
+    }
+
+    protected int reverseVerifyINodesPartial(final List<INode> inodes, final String[] names, final long[] parentIds,
+        final long[] inodeIds) throws IOException {
+      int maxIndex = (int) StatUtils.min(new double[]{inodes.size(), inodeIds.length, parentIds.length, names.length});
+      for (int i = 1; i <= maxIndex; i++) {
+        INode inode = inodes.get(inodes.size() - i);
+        boolean noChangeInInodes = inode != null && inode.getLocalName().equals(names[names.length - i]) && inode.
+            getParentId() == parentIds[parentIds.length - i] && inode.getId() == inodeIds[inodeIds.length - i];
+        if (!noChangeInInodes) {
+          return i;
+        }
+      }
+      return maxIndex + 1;
+    }
+
+    protected long[] getParentIds(long[] inodeIds) {
+      return getParentIds(inodeIds, false);
+    }
+
+    protected long[] getParentIds(long[] inodeIds, boolean partial) {
+      long[] parentIds = new long[partial ? inodeIds.length + 1 : inodeIds.length];
+      parentIds[0] = INodeDirectory.ROOT_PARENT_ID;
+      System.arraycopy(inodeIds, 0, parentIds, 1, (partial ? inodeIds.length
+          : inodeIds.length - 1));
+      return parentIds;
+    }
+
+    protected void setPartitionKey(long[] inodeIds, long parentIds[], long partitionIds[], boolean partial)
+        throws TransactionContextException, StorageException {
+      Long partId = null;
+      if (partial) {
+        if (setRandomParitionKeyEnabled && partId == null) {
+          LOG.debug("Setting Random PartitionKey");
+          partId = Math.abs(rand.nextLong());
+        }
+      } else {
+        partId = inodeIds[inodeIds.length - 1];
+      }
+      setPartitioningKey(partId);
+    }
+  }
+
+  protected class PathResolver extends CacheResolver {
+
+    @Override
+    List<INode> fetchINodes(final TransactionLockTypes.INodeLockType lockType, String path, boolean resolveLink) throws
+        IOException {
+      long[] inodeIds = Cache.getInstance().get(path);
+      if (inodeIds != null) {
+        final String[] names = INode.getPathNames(path);
+        final boolean partial = names.length > inodeIds.length;
+
+        final long[] parentIds = getParentIds(inodeIds, partial);
+        final long[] partitionIds = new long[parentIds.length];
+
+        short depth = INodeDirectory.ROOT_DIR_DEPTH;
+        partitionIds[0] = INodeDirectory.getRootDirPartitionKey();
+        for (int i = 1; i < partitionIds.length; i++) {
+          depth++;
+          partitionIds[i] = INode.calculatePartitionId(parentIds[i], names[i], depth);
+        }
+
+        setPartitionKey(inodeIds, parentIds, partitionIds, partial);
+
+        List<INode> inodes = readINodesWhileRespectingLocks(lockType, path, names,
+            parentIds, partitionIds, resolveLink);
+        if (inodes != null && !inodes.isEmpty()) {
+          final int verifiedInode = verifyINodesPartial(inodes, names,
+              parentIds, inodeIds);
+
+          int diff = inodes.size() - verifiedInode;
+          while (diff > 0) {
+            INode node = inodes.remove(inodes.size() - 1);
+            if (node != null) {
+              Cache.getInstance().delete(node);
+            }
+            diff--;
+          }
+
+          if (verifiedInode <= 1) {
+            return null;
+          }
+
+          tryResolvingTheRest(lockType, path, inodes, resolveLink);
+          return inodes;
+        }
+      }
+      return null;
+    }
+
+    INode lockInode(final TransactionLockTypes.INodeLockType lockType, long inodeId) throws IOException {
+      setINodeLockType(lockType);
+      INode targetInode = INodeUtil.getNode(inodeId, true);
+      setINodeLockType(getDefaultInodeLockType());
+      if (targetInode == null) {
+        //we throw a LeaseExpiredException as this should be called only when we try to edit an open file
+        //and the exception normaly sent if the file does not exist. This is to be compatible with apache client.
+        throw new LeaseExpiredException("No lease on inode: " + inodeId + ": File does not exist. ");
+      }
+      Cache.getInstance().set(targetInode);
+      return targetInode;
+    }
+
+    List<INode> lockInodeAndParent(final TransactionLockTypes.INodeLockType lockType, long inodeId) throws IOException {
+      List<INode> inodes = new LinkedList<>();
+      INodeIdentifier targetIdentifier = Cache.getInstance().get(inodeId);
+      if (targetIdentifier == null) {
+        INode targetInode = INodeUtil.getNode(inodeId, false);
+        if (targetInode == null) {
+          //we throw a LeaseExpiredException as this should be called only when we try to edit an open file
+          //and the exception normaly sent if the file does not exist. This is to be compatible with apache client.
+          throw new LeaseExpiredException("No lease on " + inodeId + ": File does not exist. ");
+        }
+        targetIdentifier = new INodeIdentifier(targetInode.getId(), targetInode.getParentId(), targetInode.
+            getLocalName(), targetInode.getPartitionId());
+      }
+
+      if (targetIdentifier.getInodeId() == INode.ROOT_INODE_ID) {
+        inodes.add(lockInode(lockType, inodeId));
+        return inodes;
+      }
+      setINodeLockType(TransactionLockTypes.INodeLockType.WRITE);
+      INode parentInode = INodeUtil.getNode(targetIdentifier.getPid(), true);
+      if (parentInode == null) {
+        //the cached inode does not match the one in the DB remove from cache and retry transaction
+        Cache.getInstance().delete(targetIdentifier);
+        throw new TransientStorageException("wrong inode in the cache");
+      }
+      inodes.add(parentInode);
+      Cache.getInstance().set(parentInode);
+      INode targetINode = lockInode(lockType, inodeId);
+      if (targetINode.getParentId() != parentInode.getId()) {
+        //the cached inode didn't match the one in the DB, it has now been overwriten by the one in the db, retry transaction
+        throw new TransientStorageException("wrong inode in the cache");
+      }
+      inodes.add(targetINode);
+      return inodes;
+    }
+
+    void getPathInodes(long parentId, Map<Long, INode> alreadyFetchedInodes) throws IOException {
+      if (parentId == INode.ROOT_PARENT_ID) {
+        return;
+      }
+      INodeIdentifier cur = Cache.getInstance().get(parentId);
+      if (cur == null) {
+        INode inode = INodeUtil.getNode(parentId, true);
+        if (inode == null) {
+          return;
+        }
+        Cache.getInstance().set(inode);
+        alreadyFetchedInodes.put(inode.getId(), inode);
+        cur = new INodeIdentifier(inode.getId(), inode.getParentId(), inode.getLocalName(), inode.getPartitionId());
+      }
+      Map<Long, INodeIdentifier> inodeIdentifiers = new HashMap<>();
+      while (cur != null) {
+        inodeIdentifiers.put(cur.getInodeId(), cur);
+        parentId = cur.getPid();
+        cur = Cache.getInstance().get(parentId);
+      }
+
+      for (INodeIdentifier identifier : inodeIdentifiers.values()) {
+        if (alreadyFetchedInodes.containsKey(identifier.getInodeId())) {
+          inodeIdentifiers.remove(identifier);
+        }
+      }
+
+      if (!inodeIdentifiers.isEmpty()) {
+        long[] inodeIds = new long[inodeIdentifiers.size()];
+        final String[] names = new String[inodeIdentifiers.size()];
+        final long[] parentIds = new long[inodeIdentifiers.size()];
+        final long[] partitionIds = new long[inodeIdentifiers.size()];
+
+        int i = 0;
+        for (INodeIdentifier inode : inodeIdentifiers.values()) {
+          inodeIds[i] = inode.getInodeId();
+          names[i] = inode.getName();
+          parentIds[i] = inode.getPid();
+          partitionIds[i] = inode.getPartitionId();
+          i++;
+        }
+
+        List<INode> inodesFound = find(getDefaultInodeLockType(), names, parentIds, partitionIds, true);
+        for (INode inode : inodesFound) {
+          INodeIdentifier identifier = inodeIdentifiers.get(inode.getId());
+          if (identifier != null && inode.getLocalName().equals(identifier.getName()) && inode.getParentId()
+              == identifier.getPid()) {
+            inodeIdentifiers.remove(inode.getId());
+            alreadyFetchedInodes.put(inode.getId(), inode);
+            Cache.getInstance().set(inode);
+          }
+        }
+        for (INodeIdentifier identifier : inodeIdentifiers.values()) {
+          //these identifier are in the cache but do not match the DB remove from cache and get from DB
+          Cache.getInstance().delete(identifier);
+          getPathInodes(identifier.getInodeId(), alreadyFetchedInodes);
+        }
+      }
+      if (parentId != INode.ROOT_PARENT_ID) {
+        getPathInodes(parentId, alreadyFetchedInodes);
+      }
+    }
+
+    @Override
+    List<INode> fetchINodes(final TransactionLockTypes.INodeLockType lockType, long inodeId) throws IOException {
+      List<INode> inodes = new LinkedList<>();
+      //get the inode and its parrent from the cache
+      if (TransactionLockTypes.impliesParentWriteLock(lockType)) {
+        inodes.addAll(lockInodeAndParent(lockType, inodeId));
+      } else {
+        inodes.add(lockInode(lockType, inodeId));
+      }
+
+      //we are now sure that nothing should rewrite the parents path, build it
+      Map<Long, INode> parentsMap = new HashMap<>();
+      getPathInodes(inodes.get(0).getParentId(), parentsMap);
+      INode cur = parentsMap.get(inodes.get(0).getParentId());
+      while (cur != null) {
+        inodes.add(0, cur);
+        cur = parentsMap.get(cur.getParentId());
+      }
+      return inodes;
+    }
+
+    protected void tryResolvingTheRest(final TransactionLockTypes.INodeLockType lockType, String path,
+        List<INode> inodes, boolean resolveLink)
+        throws TransactionContextException, UnresolvedPathException, StorageException {
+      int offset = inodes.size();
+
+      resolveRestOfThePath(lockType, path, inodes, resolveLink);
+      addPathINodesWithOffset(path, inodes, offset);
+    }
+
+    private void addPathINodesWithOffset(String path, List<INode> inodes, int offset) {
+      addPathINodes(path, inodes);
+      if (offset == 0) {
+        updateResolvingCache(path, inodes);
+      } else {
+        if (offset == inodes.size()) {
+          return;
+        }
+        List<INode> newInodes = inodes.subList(offset, inodes.size());
+        String[] newPath = Arrays.copyOfRange(INode.getPathNames(path), offset,
+            inodes.size());
+        updateResolvingCache(
+            Joiner.on(Path.SEPARATOR_CHAR).join(newPath), newInodes);
+      }
+    }
+
+    protected List<INode> readINodesWhileRespectingLocks(final TransactionLockTypes.INodeLockType lockType,
+        final String path,
+        final String[] names, final long[] parentIds, final long[] partitionIds, boolean resolveLink)
+        throws TransactionContextException, StorageException,
+        UnresolvedPathException {
+      int rowsToReadWithDefaultLock = names.length;
+      if (!lockType.equals(getDefaultInodeLockType())) {
+        if (lockType.equals(
+            TransactionLockTypes.INodeLockType.WRITE_ON_TARGET_AND_PARENT)) {
+          rowsToReadWithDefaultLock -= 2;
+        } else {
+          rowsToReadWithDefaultLock -= 1;
+        }
+      }
+
+      rowsToReadWithDefaultLock = Math.min(rowsToReadWithDefaultLock,
+          parentIds.length);
+
+      List<INode> inodes = null;
+      if (rowsToReadWithDefaultLock > 0) {
+        inodes = find(getDefaultInodeLockType(),
+            Arrays.copyOf(names, rowsToReadWithDefaultLock),
+            Arrays.copyOf(parentIds, rowsToReadWithDefaultLock),
+            Arrays.copyOf(partitionIds, rowsToReadWithDefaultLock), true);
+      }
+
+      if (inodes != null) {
+        for (INode inode : inodes) {
+          addLockedINodes(inode, getDefaultInodeLockType());
+        }
+      }
+
+      if (rowsToReadWithDefaultLock == names.length) {
+        return inodes;
+      }
+
+      boolean partialPath = parentIds.length < names.length;
+
+      if (inodes != null && !partialPath) {
+        resolveRestOfThePath(lockType, path, inodes, resolveLink);
+      }
+      return inodes;
+    }
+
+    protected void resolveRestOfThePath(final TransactionLockTypes.INodeLockType lockType, String path,
+        List<INode> inodes, boolean resolveLink)
+        throws StorageException, TransactionContextException,
+        UnresolvedPathException {
+      byte[][] components = INode.getPathComponents(path);
+      INode currentINode = inodes.get(inodes.size() - 1);
+      INodeResolver resolver = new INodeResolver(components, currentINode, resolveLink, true,
+          inodes.size() - 1);
+      while (resolver.hasNext()) {
+        TransactionLockTypes.INodeLockType currentINodeLock = identifyLockType(lockType, resolver.getCount() + 1,
+            components);
+        setINodeLockType(currentINodeLock);
+        currentINode = resolver.next();
+        if (currentINode != null) {
+          addLockedINodes(currentINode, currentINodeLock);
+          inodes.add(currentINode);
+        }
+      }
+    }
+  }
+
+  protected TransactionLockTypes.INodeLockType identifyLockType(final TransactionLockTypes.INodeLockType lockType,
+      int count,
+      byte[][] components) {
+    TransactionLockTypes.INodeLockType lkType;
+    if (isTarget(count, components)) {
+      lkType = lockType;
+    } else if (isParent(count, components) && TransactionLockTypes.impliesParentWriteLock(lockType)) {
+      lkType = TransactionLockTypes.INodeLockType.WRITE;
+    } else {
+      lkType = getDefaultInodeLockType();
+    }
+    return lkType;
+  }
+
+  protected boolean isTarget(int count, byte[][] components) {
+    return count == components.length - 1;
+  }
+
+  protected boolean isParent(int count, byte[][] components) {
+    return count == components.length - 2;
+  }
+
+  protected List<INode> resolveUsingCache(final TransactionLockTypes.INodeLockType lockType, long inodeId) throws
+      IOException {
+    CacheResolver cacheResolver = getCacheResolver();
+    return cacheResolver.fetchINodes(lockType, inodeId);
+  }
+
+  private CacheResolver instance = null;
+
+  protected CacheResolver getCacheResolver() {
+    if (instance == null) {
+      instance = new PathResolver();
+    }
+    return instance;
   }
 }
